# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Bank Card Management System
  description: REST API for managing bank cards, users and transfers
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
  /api/users:
    get:
      tags: [ Users ]
      summary: Get all users (ADMIN)
      operationId: getUsers
      parameters:
        - name: pageRequestDto
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/PageRequestDto"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageUserDto"
    post:
      tags: [ Users ]
      summary: Create user (ADMIN)
      operationId: addUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDto"
...
  /api/users/{id}/role:
    patch:
      tags:
        [ Users ]
      summary: Change role (ADMIN)
      operationId: changeRole
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeRoleRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDto"
  /api/users/{id}/reset-password:
    patch:
      tags: [ Users ]
      summary: Reset password (ADMIN)
      operationId: resetPassword
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: newPassword
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDto"
  /api/users/{id}/password:
    patch:
      tags: [ Users ]
      summary: Change password (USER)
      operationId: changePassword
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDto"
...
  /api/users/{id}:
    get:
      tags: [ Users ]
      summary: Get user by ID (ADMIN)
      operationId: getUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDto"
    delete:
      tags: [ Users ]
      summary: Delete a user (ADMIN)
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
...
components:
  schemas:
    CreateUserRequest:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserDto"
        password:
          type: string
          minLength: 1
      required:
        - password
    UserDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        role:
          type: string
          enum:
            - ADMIN
            - USER
        passwordHash:
          type: string
          writeOnly: true
...
    ChangeRoleRequest:
      type: object
      properties:
        role:
          type: string
          enum:
            - ADMIN
            - USER
      required:
        - role
    ChangePasswordRequest:
      type: object
      properties:
        oldPassword:
          type: string
          minLength: 1
        newPassword:
          type: string
          minLength: 1
      required:
        - newPassword
        - oldPassword
    PageRequestDto:
      type: object
      properties:
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        direction:
          type: string
          enum:
            - ASC
            - DESC
        sortBy:
          type: string
    PageUserDto:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/UserDto"
        number:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        sort:
          $ref: "#/components/schemas/SortObject"
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/PageableObject"
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: "#/components/schemas/SortObject"
        paged:
          type: boolean
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
...